feat(claw): wire org instance recover with Kilo CLI runs
Extract KiloCliRunView into a shared context-aware component so both
personal and org pages reuse the same presentation and hook dispatch.
StartKiloCliRunDialog now receives mutations as a prop and navigates
to the correct org-scoped route. Adds the org kilo-cli-run/[runId]
Next.js route following the OrganizationByPageLayout pattern.
